third_party/WebKit/LayoutTests/http/tests/linkHeader/link-preload-in-iframe.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
    var t = async_test('Makes sure that Link headers are loaded from an iframe');
    var count = 0;
    addEventListener("message", t.step_func(function(event) {
        assert_equals(event.data, "squareloaded", "FAIL: image resource was not loaded");
        ++count;
        if (count == 3) {
          t.done();
        }
        }, false));
</script>
<iframe src="resources/iframe-link-headers.php"></iframe>
